(refactor) Extract `levels/Sewer` from `levels`
